LayoutTests/http/tests/navigation/no-referrer-reset-expected.txt

CONSOLE MESSAGE: line 15: Referrer: http://127.0.0.1:8000/navigation/resources/no-referrer-reset-helper.php
CONSOLE MESSAGE: line 16: window.opener: http://127.0.0.1:8000/navigation/no-referrer-reset.html
This tests whether referrer information gets properly set and reset when "noreferrer" links are present. We do the following:
1. Open a link in a new window: referrer is sent and window.opener is sent.
2. Click a rel="noreferrer" link: referrer is null, but window.opener remains set since the link was not opened with target="_blank".
3. Click a link without rel="noreferrer": referrer is sent, but window.opener is still set. 
Start reset test
Referrer: http://127.0.0.1:8000/navigation/no-referrer-reset.html
window.opener: http://127.0.0.1:8000/navigation/no-referrer-reset.html
Referrer:
window.opener: http://127.0.0.1:8000/navigation/no-referrer-reset.html
